+++ petlog-api/src/main/java/com/ppp/api/user/dto/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.ppp.api.user.dto;
+package com.ppp.api.user.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
